Rename test.go files to exclude them from published packages
